// src/app/components/dashboard/groups/add-group/add-group.interface.ts
export interface GroupResult {
    status: any;
    message: string;
    access_token?: string;
}

export interface createGroup {
    name: string;
    userid:BigInt[];
    companyID:BigInt;
   

}

export interface User {
    id: number;
    name: string;
    lastname: string;
    email: string;
    registerDate: string;
}